TF-971 Separate binding autocomplete
(cherry picked from commit 90c3e1c8edcd3b4ecc999082a712833326205656)
